Backend OK

+++ Point.h
@@ -15,9 +15,9 @@ public:
 	void move(float x, float y);
 	void print();
 	// Calcula la distancia entre puntos
-	double getDistance(Point& p2);
+	float getDistance(Point& p2);
 	// Calcula el angulo entre puntos
-	double getAngle(Point& p2);
+	float getAngle(Point& p2);
 	// Verifica si dos puntos son muy proximos
 	bool Equal(Point& p2);
 	// Calcula un segundo punto con la distancia y el angulo de un segundo punto a este primero
